[FO-8] feat:producer_admin privileges CRUD

+++ routes/api.php
@@ -29,11 +29,11 @@ Route::prefix('v1')->group(function () {
 
         Route::get('mealCategories', 'MealCategoryController@index');
         Route::get('mealCategories/create', 'MealCategoryController@create');
-        Route::post('mealCategories', 'MealCategoryController@store');
+        Route::post('mealCategories', 'MealCategoryController@store')->middleware('role:producer_admin');
         Route::get('mealCategories/{id}', 'MealCategoryController@show');
         Route::get('mealCategories/{id}/edit', 'MealCategoryController@edit');
-        Route::put('mealCategories/{id}', 'MealCategoryController@update');
-        Route::delete('mealCategories/{id}', 'MealCategoryController@destroy');
+        Route::put('mealCategories/{id}', 'MealCategoryController@update')->middleware('role:producer_admin');
+        Route::delete('mealCategories/{id}', 'MealCategoryController@destroy')->middleware('role:producer_admin');
 
     });
 });
